+++ opensyde_tool/src/project_operations/C_PopFileTableModel.cpp
@@ -91,7 +91,7 @@ int32_t C_PopFileTableModel::rowCount(const QModelIndex & orc_Parent) const
 
    if (!orc_Parent.isValid())
    {
-      s32_Retval = this->mc_RecentFilePaths.size();
+      s32_Retval = static_cast<int32_t>(this->mc_RecentFilePaths.size());
    }
    return s32_Retval;
 }
